\name{OptimalTrading-package}
\alias{OptimalTrading-package}
\alias{OptimalTrading}
\docType{package}
\title{
Analyze Historic Data and Compute Optimal Trading Strategies
}
\description{
Based on historical prices, backtest and find what the optimal trading strategy
would have been.  Benchmark your own trading strategy with the optimal one.  Use
association analysis to find any rules among a trading strategy.
}
\details{
\tabular{ll}{
Package: \tab OptimalTrading\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-22\cr
License: \tab Private license, contact jeff.ct.wong@stanford.edu for permissions\cr
LazyLoad: \tab yes\cr
Depends: \tab arules, lpSolve, lpSolveAPI, quadprog, Heatplus\cr
}
}
\author{
Jeffrey Wong <jeff.ct.wong@stanford.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
